pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-maintenance"
version = "0.1.0"
description = "CLI tool for maintaining and preparing local skill libraries for coding agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "lazarosvarvatis", email = "lazarosvarvatis@users.noreply.github.com" }]
keywords = ["cli", "agent", "skills", "developer-tools", "llm", "coding-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
embeddings = [
    "sentence-transformers>=3.0",
]
llm = [
    "anthropic>=0.40",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/lazarosvarvatis/agent-maintenance"
Issues = "https://github.com/lazarosvarvatis/agent-maintenance/issues"

[project.scripts]
agent-maintenance = "agent_maintenance.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_maintenance"]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=agent_maintenance --cov-report=term-missing -v"
